+++ src/components/comments/comments_interfaces.ts
@@ -1,0 +1,13 @@
+
+
+interface INewComment {
+    userId: number;
+    postId: number;
+    content: string;
+}
+
+interface IComment extends INewComment {
+    id: number;
+}
+
+export {IComment, INewComment};
